+++ internal/services/blog/blog.go
@@ -1,0 +1,168 @@
+package blog
+
+import (
+	"context"
+	"fmt"
+	blogpb "todo-app-grpc/proto"
+
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+)
+
+type BlogServiceServer struct {
+	db *mongo.Collection
+}
+
+type Blog interface {
+	CreateBlog(context.Context, *blogpb.CreateBlogReq) (*blogpb.CreateBlogRes, error)
+	ReadBlog(context.Context, *blogpb.ReadBlogReq) (*blogpb.ReadBlogRes, error)
+	UpdateBlog(context.Context, *blogpb.UpdateBlogReq) (*blogpb.UpdateBlogRes, error)
+	DeleteBlog(context.Context, *blogpb.DeleteBLogReq) (*blogpb.DeleteBlogRes, error)
+	ListBlogs(*blogpb.ListBlogsReq, blogpb.BlogService_ListBlogsServer) error
+}
+
+func New(db *mongo.Collection) *BlogServiceServer {
+	return &BlogServiceServer{
+		db: db,
+	}
+}
+
+func (s *BlogServiceServer) CreateBlog(ctx context.Context, req *blogpb.CreateBlogReq) (*blogpb.CreateBlogRes, error) {
+	blog := req.GetBlog()
+
+	data := BlogItem{
+		AuthorID: blog.GetAuthorId(),
+		Title:    blog.GetTitle(),
+		Content:  blog.GetContent(),
+	}
+
+	// Insert the data into the database, result contains the newly generated Object ID for the new document
+	result, err := s.db.InsertOne(ctx, data)
+	if err != nil {
+		return nil, status.Errorf(
+			codes.Internal,
+			fmt.Sprintf("internal error: %v", err),
+		)
+	}
+	oid := result.InsertedID.(primitive.ObjectID)
+	blog.Id = oid.Hex()
+	return &blogpb.CreateBlogRes{Blog: blog}, nil
+}
+
+func (s *BlogServiceServer) ReadBlog(ctx context.Context, req *blogpb.ReadBlogReq) (*blogpb.ReadBlogRes, error) {
+	// convert string id (from proto) to mongoDB ObjectId
+	oid, err := primitive.ObjectIDFromHex(req.GetId())
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, fmt.Sprintf("could not to convert to objectID: %v", err))
+	}
+	result := s.db.FindOne(ctx, bson.M{"_id": oid})
+	data := BlogItem{}
+	if err := result.Decode(&data); err != nil {
+		return nil, status.Errorf(codes.NotFound, fmt.Sprintf("Could not find blog with Object ID %s: %v", req.GetId(), err))
+	}
+	responce := &blogpb.ReadBlogRes{
+		Blog: &blogpb.Blog{
+			Id:       oid.Hex(),
+			AuthorId: data.AuthorID,
+			Title:    data.Title,
+			Content:  data.Content,
+		},
+	}
+	return responce, nil
+
+}
+
+func (s *BlogServiceServer) DeleteBlog(ctx context.Context, req *blogpb.DeleteBLogReq) (*blogpb.DeleteBlogRes, error) {
+	oid, err := primitive.ObjectIDFromHex(req.GetId())
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, fmt.Sprintf("Could not convert to ObjectID: %v", err))
+	}
+	// DeleteOne returns DeleteResult which is a struct containing the amount of deleted docs (in this case only 1 always)
+	// So we return a boolean instead
+	_, err = s.db.DeleteOne(ctx, bson.M{"_id": oid})
+	if err != nil {
+		return nil, status.Errorf(codes.NotFound, fmt.Sprintf("Could not find/delete blog with id %s: %v", req.GetId(), err))
+	}
+
+	return &blogpb.DeleteBlogRes{
+		Sucsess: true,
+	}, nil
+}
+
+func (s *BlogServiceServer) UpdateBlog(ctx context.Context, req *blogpb.UpdateBlogReq) (*blogpb.UpdateBlogRes, error) {
+	blog := req.GetBlog()
+
+	// Convert the Id string to a MongoDB ObjectId
+	oid, err := primitive.ObjectIDFromHex(blog.GetId())
+	if err != nil {
+		return nil, status.Errorf(
+			codes.InvalidArgument,
+			fmt.Sprintf("Could not convert the supplied blog id to mongoDB ObjectID: %v", err),
+		)
+	}
+	update := bson.M{
+		"author_id": blog.GetAuthorId(),
+		"title":     blog.GetTitle(),
+		"content":   blog.GetContent(),
+	}
+
+	// Convert the oid into an unordered bson document to search by id
+	filter := bson.M{"_id": oid}
+
+	// Result is the BSON encoded result
+	// To return the updated document instead of original we have to add options.
+	result := s.db.FindOneAndUpdate(ctx, filter, bson.M{"set": update}, options.FindOneAndUpdate().SetReturnDocument(1))
+
+	// Decode result and write it to 'decoded'
+	decoded := BlogItem{}
+	err = result.Decode(&decoded)
+	if err != nil {
+		return nil, status.Errorf(
+			codes.NotFound,
+			fmt.Sprintf("Could not find blog with supplied ID: %v", err),
+		)
+	}
+	return &blogpb.UpdateBlogRes{
+		Blog: &blogpb.Blog{
+			Id:       decoded.ID.Hex(),
+			AuthorId: decoded.AuthorID,
+			Title:    decoded.Title,
+			Content:  decoded.Content,
+		},
+	}, nil
+}
+func (s *BlogServiceServer) ListBlogs(req *blogpb.ListBlogsReq, stream blogpb.BlogService_ListBlogsServer) error {
+	// Initiate a BlogItem type to write decoded data to
+	data := &BlogItem{}
+	// collection.Find returns a cursor for our (empty) query
+	cursor, err := s.db.Find(context.Background(), bson.M{})
+	if err != nil {
+		return status.Errorf(codes.Internal, fmt.Sprintf("Unknown internal error: %v", err))
+	}
+	defer cursor.Close(context.Background())
+	// cursor.Next() returns a boolean, if false there are no more items and loop will break
+	for cursor.Next(context.Background()) {
+		// Decode the data at the current pointer and write it to data
+		err := cursor.Decode(data)
+		if err != nil {
+			return status.Errorf(codes.Unavailable, fmt.Sprintf("Could not decode data: %v", err))
+		}
+		// If no error is found send blog over stream
+		stream.Send(&blogpb.ListBlogRes{
+			Blog: &blogpb.Blog{
+				Id:       data.ID.Hex(),
+				AuthorId: data.AuthorID,
+				Content:  data.Content,
+				Title:    data.Title,
+			},
+		})
+	}
+	if err := cursor.Err(); err != nil {
+		return status.Errorf(codes.Internal, fmt.Sprintf("Unknown cursor error: %v", err))
+	}
+	return nil
+}
