+++ internal/services/blog/model.go
@@ -1,0 +1,10 @@
+package blog
+
+import "go.mongodb.org/mongo-driver/bson/primitive"
+
+type BlogItem struct {
+	ID       primitive.ObjectID `bson:"_id,omitempty" `
+	AuthorID string             ` bson: "author_id"`
+	Content  string             ` bson: "content"`
+	Title    string             `bson:"title"`
+}
