+++ server/main.go
@@ -8,166 +8,12 @@ import (
 	"os"
 	"os/signal"
 
+	db "todo-app-grpc/internal/services/DB"
+	"todo-app-grpc/internal/services/blog"
 	blogpb "todo-app-grpc/proto"
 
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/bson/primitive"
-	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
-
-type BlogServiceServer struct {
-}
-
-var db *mongo.Client
-var blogdb *mongo.Collection
-var mongoCtx context.Context
-
-type BlogItem struct {
-	ID       primitive.ObjectID `bson:"_id,omitempty" `
-	AuthorID string             ` bson: "author_id"`
-	Content  string             ` bson: "content"`
-	Title    string             `bson:"title"`
-}
-
-func (s *BlogServiceServer) CreateBlog(ctx context.Context, req *blogpb.CreateBlogReq) (*blogpb.CreateBlogRes, error) {
-	blog := req.GetBlog()
-
-	data := BlogItem{
-		AuthorID: blog.GetAuthorId(),
-		Title:    blog.GetTitle(),
-		Content:  blog.GetContent(),
-	}
-
-	// Insert the data into the database, result contains the newly generated Object ID for the new document
-	result, err := blogdb.InsertOne(mongoCtx, data)
-	if err != nil {
-		return nil, status.Errorf(
-			codes.Internal,
-			fmt.Sprintf("internal error: %v", err),
-		)
-	}
-	oid := result.InsertedID.(primitive.ObjectID)
-	blog.Id = oid.Hex()
-	return &blogpb.CreateBlogRes{Blog: blog}, nil
-}
-
-func (s *BlogServiceServer) ReadBlog(ctx context.Context, req *blogpb.ReadBlogReq) (*blogpb.ReadBlogRes, error) {
-	// convert string id (from proto) to mongoDB ObjectId
-	oid, err := primitive.ObjectIDFromHex(req.GetId())
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, fmt.Sprintf("could not to convert to objectID: %v", err))
-	}
-	result := blogdb.FindOne(ctx, bson.M{"_id": oid})
-	data := BlogItem{}
-	if err := result.Decode(&data); err != nil {
-		return nil, status.Errorf(codes.NotFound, fmt.Sprintf("Could not find blog with Object ID %s: %v", req.GetId(), err))
-	}
-	responce := &blogpb.ReadBlogRes{
-		Blog: &blogpb.Blog{
-			Id:       oid.Hex(),
-			AuthorId: data.AuthorID,
-			Title:    data.Title,
-			Content:  data.Content,
-		},
-	}
-	return responce, nil
-
-}
-
-func (s *BlogServiceServer) DeleteBlog(ctx context.Context, req *blogpb.DeleteBLogReq) (*blogpb.DeleteBlogRes, error) {
-	oid, err := primitive.ObjectIDFromHex(req.GetId())
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, fmt.Sprintf("Could not convert to ObjectID: %v", err))
-	}
-	// DeleteOne returns DeleteResult which is a struct containing the amount of deleted docs (in this case only 1 always)
-	// So we return a boolean instead
-	_, err = blogdb.DeleteOne(ctx, bson.M{"_id": oid})
-	if err != nil {
-		return nil, status.Errorf(codes.NotFound, fmt.Sprintf("Could not find/delete blog with id %s: %v", req.GetId(), err))
-	}
-
-	return &blogpb.DeleteBlogRes{
-		Sucsess: true,
-	}, nil
-}
-
-func (s *BlogServiceServer) UpdateBlog(ctx context.Context, req *blogpb.UpdateBlogReq) (*blogpb.UpdateBlogRes, error) {
-	blog := req.GetBlog()
-
-	// Convert the Id string to a MongoDB ObjectId
-	oid, err := primitive.ObjectIDFromHex(blog.GetId())
-	if err != nil {
-		return nil, status.Errorf(
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not convert the supplied blog id to mongoDB ObjectID: %v", err),
-		)
-	}
-	update := bson.M{
-		"author_id": blog.GetAuthorId(),
-		"title":     blog.GetTitle(),
-		"content":   blog.GetContent(),
-	}
-
-	// Convert the oid into an unordered bson document to search by id
-	filter := bson.M{"_id": oid}
-
-	// Result is the BSON encoded result
-	// To return the updated document instead of original we have to add options.
-	result := blogdb.FindOneAndUpdate(ctx, filter, bson.M{"set": update}, options.FindOneAndUpdate().SetReturnDocument(1))
-
-	// Decode result and write it to 'decoded'
-	decoded := BlogItem{}
-	err = result.Decode(&decoded)
-	if err != nil {
-		return nil, status.Errorf(
-			codes.NotFound,
-			fmt.Sprintf("Could not find blog with supplied ID: %v", err),
-		)
-	}
-	return &blogpb.UpdateBlogRes{
-		Blog: &blogpb.Blog{
-			Id:       decoded.ID.Hex(),
-			AuthorId: decoded.AuthorID,
-			Title:    decoded.Title,
-			Content:  decoded.Content,
-		},
-	}, nil
-}
-func (s *BlogServiceServer) ListBlogs(req *blogpb.ListBlogsReq, stream blogpb.BlogService_ListBlogsServer) error {
-	// Initiate a BlogItem type to write decoded data to
-	data := &BlogItem{}
-	// collection.Find returns a cursor for our (empty) query
-	cursor, err := blogdb.Find(context.Background(), bson.M{})
-	if err != nil {
-		return status.Errorf(codes.Internal, fmt.Sprintf("Unknown internal error: %v", err))
-	}
-	defer cursor.Close(context.Background())
-	// cursor.Next() returns a boolean, if false there are no more items and loop will break
-	for cursor.Next(context.Background()) {
-		// Decode the data at the current pointer and write it to data
-		err := cursor.Decode(data)
-		if err != nil {
-			return status.Errorf(codes.Unavailable, fmt.Sprintf("Could not decode data: %v", err))
-		}
-		// If no error is found send blog over stream
-		stream.Send(&blogpb.ListBlogRes{
-			Blog: &blogpb.Blog{
-				Id:       data.ID.Hex(),
-				AuthorId: data.AuthorID,
-				Content:  data.Content,
-				Title:    data.Title,
-			},
-		})
-	}
-	if err := cursor.Err(); err != nil {
-		return status.Errorf(codes.Internal, fmt.Sprintf("Unknown cursor error: %v", err))
-	}
-	return nil
-}
 
 func main() {
 	log.SetFlags(log.LstdFlags | log.Lshortfile)
@@ -179,27 +25,13 @@ func main() {
 	}
 	opts := []grpc.ServerOption{}
 	s := grpc.NewServer(opts...)
-	srv := &BlogServiceServer{}
+	mongoCtx := context.Background()
+	db := db.New(mongoCtx)
+	srv := blog.New(db.Database("mydb").Collection("blog"))
 
 	blogpb.RegisterBlogServiceServer(s, srv)
 	// Initialize MongoDb client
 	fmt.Println("Connecting to mongoDB...")
-
-	mongoCtx = context.Background()
-	// Connect takes in a context and options, the connection URI is the only option we pass for now
-	db, err = mongo.Connect(mongoCtx, options.Client().ApplyURI("mongodb://localhost:27017"))
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = db.Ping(mongoCtx, nil)
-	if err != nil {
-		log.Fatalf("failed to connect to mongoDB: %v\n", err)
-	} else {
-		fmt.Println("Connceted to mondoDB")
-	}
-
-	blogdb = db.Database("mydb").Collection("blog")
 
 	go func() {
 		if err := s.Serve(listener); err != nil {
@@ -215,12 +47,19 @@ func main() {
 	signal.Notify(c, os.Interrupt)
 
 	// Block main routine until a signal is received
+
 	// As long as user doesn't press CTRL+C a message is not passed and our main routine keeps running
 	<-c
+
 	fmt.Println("\nStopping the server... ")
+
 	s.Stop()
+
 	listener.Close()
+
 	fmt.Println("Closing MongoDB conncetion")
+
 	db.Disconnect(mongoCtx)
+
 	fmt.Println("Done")
 }
